feat: US-002 - Move email module files to libs
