+++ libs/email/interfaces/email-provider.interface.ts
