+++ libs/email/interfaces/send-email-options.interface.ts
